Refactor authorization scheme
Details
-------

Rewrite permission middleware to a more simple model where endpoints are
permitted by the combination of role & route. Where that model is not
applicable (all users are able to list Clients, for example),
authorization is handled by simple filters on queries to the DB. For
example, a Client will only see themselves returned in an index query
while a Superadmin will see every Client. If a user requests something
to which they are allowed access to the route, but not the actual
resource, they'll get a 404 rather than a 403. A Client attempting to
get a single other client (e.g. /api/clients/999) will return a 404.

This greatly simplifies how authz works and I think the slight loss of
fidelity is worth it.