Add created_at/updated_at to User, Org, & Request
Summary:

- User (+created_at, updated -> updated_at)
- Org (+created_at, +updated_at)
- Request (+updated_at)                      [created_at already exists]

Details:

These columns are DB-managed with a trigger around INSERT & UPDATE. User
code cannot set or modify these columns because a trigger will silently
overwrite any provided value. This might have slightly surprising
behavior -- API calls effectively can't modify these columns, or not
modify in the case of updated_at.

The benefit is that nothing has to be done in client code about these
columns. Most code stays the same. One change is making the types of
these columns optional because values do not have to be supplied to the
DB. The other change is that updated has been renamed to updated_at on
"user" for consistency.